pyproject.toml
[project]
name = "marl-uav-wind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "plotly>=6.1.0",
    "scipy>=1.15.3",
    "tqdm>=4.67.1",
]

